// packages/ui/src/selectors/PeriodSelector/components/PeriodSelect/interfaces/props.ts
export interface PeriodSelectProps {
    singleSelection?: boolean;
    excludedPeriodTypes: Array<string>;
    onSelect?: ({items}: { items: Array<string> }) => void;
    selectedPeriods?: Array<string>;
    excludeRelativePeriods?: boolean;
    excludeFixedPeriods?: boolean;
    allowFuturePeriods?: boolean;
}